Allow subcommands to implement TextUnmarshaler
In some corner cases you want to have custom parsingfor the whole
subcommand instead of just a single arg/flag. This change allows a
subcommand to define it's own implementation using the
encoding.TextUnmarshaler interface.

The limitation of this approach is that once it finds the subcommand
that has a custom implementation, it passes the rest of the args to it
and in the case that a parent command has flags, they need to come
before the subcommand.

For example:
  * parent --verbose child --what --ever arg1 arg2 (works as expected)
  * parent child --what --ever arg1 arg2 --verbose (the verbose flag
    will be passed to the custom implementation of child)